Split the plugin module
