+++ plugin/bungee/src/main/resources/bungee.yml
