+++ plugin/spigot/src/main/java/org/dragonet/plugin/bukkit/BedrockPacketProcessorRegister.java
@@ -1,6 +1,7 @@
 package org.dragonet.plugin.bukkit;
 
 import org.dragonet.plugin.bukkit.processors.ContainerCloseProcessor;
+import org.dragonet.plugin.bukkit.processors.InventoryTransactionProcessor;
 import org.dragonet.plugin.bukkit.processors.ModalFormResponseProcessor;
 import org.dragonet.protocol.PEPacket;
 import org.dragonet.protocol.packets.ContainerClosePacket;
@@ -9,7 +10,6 @@ import org.dragonet.protocol.packets.ModalFormResponsePacket;
 
 import java.util.HashMap;
 import java.util.Map;
-import org.dragonet.plugin.bukkit.processors.InventoryTransactionProcessor;
 
 public final class BedrockPacketProcessorRegister {
