+++ plugin/spigot/src/main/java/org/dragonet/plugin/bukkit/processors/ContainerCloseProcessor.java
