+++ plugin/spigot/src/main/java/org/dragonet/plugin/dpaddon/DPAddonBukkit.java
@@ -10,16 +10,16 @@
  *
  * @author The Dragonet Team
  */
-package org.dragonet.plugin.plugin.dpaddon;
+package org.dragonet.plugin.dpaddon;
 
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.plugin.java.JavaPlugin;
+import org.dragonet.common.utilities.BinaryStream;
 import org.dragonet.plugin.bukkit.BedrockPlayer;
 import org.dragonet.plugin.bukkit.DPPluginMessageListener;
-import org.dragonet.common.utilities.BinaryStream;
 
 import java.util.HashSet;
 import java.util.Set;
