+++ plugin/spigot/src/main/resources/plugin.yml
